pyproject.toml
[project]
name = "flask-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.3",
    "apiflask>=2.4.0",
    "bcrypt>=4.3.0",
    "bpython>=0.25",
    "flask>=3.1.1",
    "flask-admin[all]==2.0.0a4",
    "flask-apps>=0.0.2",
    "flask-babel>=4.0.0",
    "flask-login>=0.6.3",
    "flask-mailgun",
    "flask-mailman>=1.1.1",
    "flask-security-too>=5.6.2",
    "flask-security[common,fsqla,mfa]>=5.6.2",
    "flask-shell-ipython>=0.5.3",
    "flask-typed>=0.2.3",
    "inflection>=0.5.1",
    "mailgun>=1.1.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "sqlalchemy-utils>=0.41.2",
    "sqlmodel>=0.0.24",
    "uuid>=1.30",
    "watchdog>=6.0.0",
    "wtforms==3.0.1",
]

[dependency-groups]
dev = [
    "faker>=37.5.3",
    "ipython>=9.4.0",
    "pytap>=0.1.1",
    "pytest>=8.4.1",
    "pytest-flask>=1.3.0",
    "pytest-tap>=3.5",
    "sqlalchemy-model-faker>=0.0.5",
]
